check if swap is present

+++ .scripts/masternode_install.sh
@@ -72,15 +72,21 @@ function install_packages() {
 }
 
 function swaphack() { 
-	# needed because ant servers are ants
-	rm -f /var/swap.img
-	dd if=/dev/zero of=/var/swap.img bs=1024k count=${MNODE_SWAPSIZE}
-	chmod 0600 /var/swap.img
-	mkswap /var/swap.img
-	swapon /var/swap.img
-	echo '/var/swap.img none swap sw 0 0' | tee -a /etc/fstab
-	echo 'vm.swappiness=10' | tee -a /etc/sysctl.conf
-	echo 'vm.vfs_cache_pressure=50' | tee -a /etc/sysctl.conf	
+	#check if swap is available
+	if free | awk '/^Swap:/ {exit !$2}'; then
+		echo "Already have swap"
+	else
+		echo "No swap"
+		# needed because ant servers are ants
+		rm -f /var/swap.img
+		dd if=/dev/zero of=/var/swap.img bs=1024k count=${MNODE_SWAPSIZE}
+		chmod 0600 /var/swap.img
+		mkswap /var/swap.img
+		swapon /var/swap.img
+		echo '/var/swap.img none swap sw 0 0' | tee -a /etc/fstab
+		echo 'vm.swappiness=10' | tee -a /etc/sysctl.conf
+		echo 'vm.vfs_cache_pressure=50' | tee -a /etc/sysctl.conf		
+	fi
 }
 
 function build_mn_from_source() {
